backend/requirements.txt
Flask
psycopg2-binary
flask_jwt_extended
flask_socketio
flask_cors
flask_bcrypt
cryptography